chore: restructure localstorev2 pkg (#3777)
